Move BaseFragment into common package
