+++ app/src/main/java/org/gdg/frisbee/android/about/GetInvolvedFragment.java
@@ -22,7 +22,7 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import org.gdg.frisbee.android.R;
-import org.gdg.frisbee.android.fragment.BaseFragment;
+import org.gdg.frisbee.android.common.BaseFragment;
 
 import butterknife.ButterKnife;
