+++ app/src/main/java/org/gdg/frisbee/android/fragment/GdgListFragment.java
@@ -29,6 +29,7 @@ import android.widget.ListAdapter;
 import android.widget.ListView;
 
 import org.gdg.frisbee.android.R;
+import org.gdg.frisbee.android.common.BaseFragment;
 
 import butterknife.ButterKnife;
 import butterknife.InjectView;
